+++ src/main/java/org/homework/springhomework003/service/implement/VenueServiceImpl.java
@@ -1,5 +1,6 @@
 package org.homework.springhomework003.service.implement;
 
+import org.homework.springhomework003.exception.NotFoundException;
 import org.homework.springhomework003.model.entity.Venue;
 import org.homework.springhomework003.model.dto.request.VenueRequest;
 import org.homework.springhomework003.repository.VenueRepository;
@@ -18,16 +19,28 @@ public class VenueServiceImpl implements VenueService {
 
     @Override
     public Venue getVenueById(Integer id) {
-        return venueRepository.getVenueById(id);
+        Venue venue = venueRepository.getVenueById(id);
+        if (venue == null) {
+            throw new NotFoundException("Venue id [" + id + "] not found");
+        }
+        return venue;
     }
 
     @Override
     public Venue updateVenueById(VenueRequest venueRequest, Integer id) {
+        Venue existingVenue = venueRepository.getVenueById(id);
+        if (existingVenue == null) {
+            throw new NotFoundException("Venue id [" + id + "] not found");
+        }
         return venueRepository.updateVenueById(id, venueRequest);
     }
 
     @Override
     public Venue deleteVenueById(Integer id) {
+        Venue existingVenue = venueRepository.getVenueById(id);
+        if (existingVenue == null) {
+            throw new NotFoundException("Venue id [" + id + "] not found");
+        }
         return venueRepository.deleteVenueById(id);
     }
 
@@ -35,7 +48,6 @@ public class VenueServiceImpl implements VenueService {
     public List<Venue> getAllVenue(Integer page, Integer size) {
         return venueRepository.getAllVenues(page, size);
     }
-
 
     @Override
     public Venue addVenue(VenueRequest venueRequest) {
